requirements.txt
pandas==1.3.3
Pillow==9.0.0
plotly==5.5.0
pymongo==4.0.1
streamlit==1.3.1
tensorflow==2.4.3
pymongo[srv]>=3.11.0
